refactor to get everything building

+++ include/substratum_defs.h
@@ -26,5 +26,5 @@ typedef u8 bool;
 #define false 0
 #define true 1
 
-#define MACHINE_REGISTER_SIZE_BYTES ((size_t)8)
+#define MACHINE_REGISTER_SIZE_BYTES 8
 #define MACHINE_REGISTER_COUNT 32
